Remove old InvalidRequestException, rename FooInvalidRequestException (#1246)
