+++ src/FooRequest/YamlFieldValidator.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\FooRequest;
 
-use App\Exception\FooInvalidRequestException;
+use App\Exception\InvalidRequestException;
 use App\FooRequest\Field\Field;
 use SmartAssert\YamlFile\Validator\ContentValidator;
 
@@ -18,18 +18,18 @@ readonly class YamlFieldValidator
     /**
      * @return non-empty-string
      *
-     * @throws FooInvalidRequestException
+     * @throws InvalidRequestException
      */
     public function validate(Field $field): string
     {
         $content = $field->getValue();
         if ('' === $content) {
-            throw new FooInvalidRequestException('invalid_request_field', $field, 'empty');
+            throw new InvalidRequestException('invalid_request_field', $field, 'empty');
         }
 
         $validation = $this->yamlContentValidator->validate($content);
         if (false === $validation->isValid()) {
-            throw new FooInvalidRequestException('invalid_request_field', $field, 'invalid');
+            throw new InvalidRequestException('invalid_request_field', $field, 'invalid');
         }
 
         return $content;
