+++ src/FooRequest/YamlFilenameFieldValidator.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\FooRequest;
 
-use App\Exception\FooInvalidRequestException;
+use App\Exception\InvalidRequestException;
 use App\FooRequest\Field\YamlFilenameField;
 use SmartAssert\YamlFile\Filename as YamlFilename;
 use SmartAssert\YamlFile\Validator\YamlFilenameValidator;
@@ -17,7 +17,7 @@ readonly class YamlFilenameFieldValidator
     }
 
     /**
-     * @throws FooInvalidRequestException
+     * @throws InvalidRequestException
      */
     public function validate(YamlFilenameField $field): YamlFilename
     {
@@ -25,7 +25,7 @@ readonly class YamlFilenameFieldValidator
         $validation = $this->yamlFilenameValidator->validate($filename);
 
         if (false === $validation->isValid()) {
-            throw new FooInvalidRequestException('invalid_request_field', $field, 'invalid');
+            throw new InvalidRequestException('invalid_request_field', $field, 'invalid');
         }
 
         return $filename;
